requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
Click==7.0
colorama==0.4.3
importlib-metadata==1.4.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.1.0
packaging==20.0
pluggy==0.13.1
py==1.8.1
PyAutoIt==0.6.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.3
selenium==3.141.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.7
wcwidth==0.1.8
wincertstore==0.2
wrapt==1.11.2
zipp==1.0.0
